# dist/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Modo Challenge",
        "description": "Api de movimientos de cuentas bancarias",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
            "url": "/api/",
            "description": "Base url para acceder desde heroku a todos los endpoint"
        }],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/card": {
            "get": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Buscar todas las tarjetas",
                "parameters": [{
                        "in": "header",
                        "name": "x-token",
                        "required": true,
                        "description": "token",
                        "schema": {
                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDQ2NGY1NGUyMjRjMDJhMThlZWQ5MmYiLCJpYXQiOjE2MTUyMjQ3MTAsImV4cCI6MTYxNTI2NzkxMH0.46U86skzXpg5d-X6xuAM_qucUAzmtrrEqxJsRwtudQY",
                            "$ref": "#/definitions/id"
                        }
                    }],
                "responses": {
                    "200": {
                        "description": "movimiento por id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "number_id": {
                                                        "type": "number",
                                                        "example": 2342352323
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "Débito"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Mastercard"
                                                    }
                                                }
                                            }
                                        },
                                        "$ref": "#/definitions/user"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sin resultados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "no se encontraron resultados"
                                        }
                                    },
                                    "$ref": "#/definitions/id"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card/": {
            "post": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Crear tarjeta",
                "parameters": [{
                        "in": "header",
                        "name": "x-token",
                        "required": true,
                        "description": "token",
                        "schema": {
                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDQ2NGY1NGUyMjRjMDJhMThlZWQ5MmYiLCJpYXQiOjE2MTUyMjQ3MTAsImV4cCI6MTYxNTI2NzkxMH0.46U86skzXpg5d-X6xuAM_qucUAzmtrrEqxJsRwtudQY",
                            "$ref": "#/definitions/id"
                        }
                    }],
                "requestBody": {
                    "description": "Creación de tarjeta",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "number_id": {
                                        "type": "integer",
                                        "example": "2342423423"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Débito"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Visa"
                                    }
                                },
                                "$ref": "#/definitions/user"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movimientos de usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "El movimiento se ha creado  correctamente!!"
                                        }
                                    },
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sin resultados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "no se encontraron resultados"
                                        }
                                    },
                                    "$ref": "#/definitions/id"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card/{number_id}": {
            "delete": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Borrar tarjeta por id",
                "parameters": [{
                        "in": "header",
                        "name": "x-token",
                        "required": true,
                        "description": "token",
                        "schema": {
                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDQ2NGY1NGUyMjRjMDJhMThlZWQ5MmYiLCJpYXQiOjE2MTUyMjQ3MTAsImV4cCI6MTYxNTI2NzkxMH0.46U86skzXpg5d-X6xuAM_qucUAzmtrrEqxJsRwtudQY",
                            "$ref": "#/definitions/id"
                        }
                    }],
                "responses": {
                    "200": {
                        "description": "Tarjetas de usuario creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "El borrado fue realizado exitosamente!!"
                                        }
                                    },
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sin resultados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "El borrado no se pudo realizar"
                                        }
                                    },
                                    "$ref": "#/definitions/id"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card/modificar/{number_id}": {
            "put": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Actualizar tarjeta por Id",
                "parameters": [{
                        "in": "header",
                        "name": "x-token",
                        "required": true,
                        "description": "token",
                        "schema": {
                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MDQ2NGY1NGUyMjRjMDJhMThlZWQ5MmYiLCJpYXQiOjE2MTUyMjQ3MTAsImV4cCI6MTYxNTI2NzkxMH0.46U86skzXpg5d-X6xuAM_qucUAzmtrrEqxJsRwtudQY",
                            "$ref": "#/definitions/id"
                        }
                    }],
                "responses": {
                    "200": {
                        "description": "Movimientos de usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "El actualizado de la tarjeta fue realizado exitosamente!!"
                                        }
                                    },
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sin resultados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "El actualizado de la tarjeta no se pudo realizar"
                                        }
                                    },
                                    "$ref": "#/definitions/id"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}